feat(spec): Fractional stoichiometry across all bindings (gt-1e96)
Relax StoichiometryEntry.stoichiometry from `integer ≥ 1` to positive
finite `number` so atmospheric-chemistry mechanisms with non-integer
product yields (e.g. ISOP+O3 → 0.87 CH2O + 1.86 CH3O2 + 0.06 HO2 +
0.05 CO, CH3O2+CH3O2 → 2.0 CH2O + 0.8 HO2) round-trip losslessly
rather than requiring a lossy shadow-reaction split. Integer
coefficients remain valid and byte-identical across parse/re-emit.

Schema + all four embedded copies updated; parsers in Julia, Python,
Rust, Go, TypeScript now accept fractional values and serializers
preserve int-vs-float distinction so existing integer fixtures
retain their canonical form. SolidJS editor display adjusted.

Adds tests/valid/fractional_stoichiometry.esm (SuperFast-like
mechanism with 3+ fractional-stoich reactions) to the conformance
round-trip manifest, plus a zero-coefficient rejection fixture, and
per-binding round-trip tests (Rust, Go, Python).

+++ packages/EarthSciSerialization.jl/src/flatten.jl
@@ -254,7 +254,7 @@ function lower_reactions_to_equations(reactions::Vector{Reaction},
 
     n_species = length(species_names)
     n_rxns = length(reactions)
-    S = zeros(Int, n_species, n_rxns)
+    S = zeros(Float64, n_species, n_rxns)
 
     for (j, rxn) in enumerate(reactions)
         substrates = getfield(rxn, :substrates)
